+++ Demo/Application/Features/PayPal - Pay Later/BraintreeDemoPayPalPayLaterViewController.swift
@@ -27,7 +27,7 @@ class BraintreeDemoPayPalPayLaterViewController: BraintreeDemoPaymentButtonBaseV
         request.offerPayLater = true
         request.activeWindow = self.view.window
         
-        client.tokenizePayPalAccount(with: request) { nonce, error in
+        client.tokenize(request) { nonce, error in
             sender.isEnabled = true
             
             guard let nonce = nonce else {